Refactored classes, structs, and inline functions into header file.
