+++ streamcluster/src-rl/streamcluster.hpp
@@ -1,0 +1,126 @@
+#ifndef STREAMCLUSTER_H
+#define STREAMCLUSTER_H
+
+#include <stdio.h>
+#include <iostream>
+#include <fstream>
+#include <stdlib.h>
+#include <cstring>
+#include <vector>
+#include <assert.h>
+#include <math.h>
+#include <sys/resource.h>
+#include <climits>
+
+/* this structure represents a point */
+/* these will be passed around to avoid copying coordinates */
+struct Point 
+{
+  float     weight  = 1.0;
+  float     *coord  = nullptr;
+  long      assign  = 0;  /* number of point where this one is assigned */
+  float     cost    = 0.0;  /* cost of that assignment, weight*distance */
+};
+
+/* this is the array of points */
+struct Points 
+{
+    Points( long n, long d, std::size_t count ) : num( n ), dim( d )
+    {
+        p.reserve( count );
+    }
+
+    long  num; /* number of points; may not be N if this is a sample */
+    int   dim;  /* dimensionality */
+    std::vector< Point > p; /* the array itself */
+};
+
+struct pkmedian_arg_t
+{
+  Points* points;
+  long kmin;
+  long kmax;
+  long* kfinal;
+  int pid;
+  pthread_barrier_t* barrier;
+};
+
+class PStream {
+public:
+  PStream() = default;
+  virtual ~PStream() = default;
+  
+  virtual size_t read( float* dest, int dim, int num ) = 0;
+  virtual int ferror() = 0;
+  virtual int feof() = 0;
+};
+
+//synthetic stream
+class SimStream : public PStream 
+{
+public:
+  SimStream(long n_ ) : PStream(), n( n_ ) {};
+
+  size_t read( float* dest, int dim, int num ) 
+  {
+    size_t count = 0;
+    for( int i = 0; i < num && n > 0; i++ ) 
+    {
+      for( int k = 0; k < dim; k++ ) 
+      {
+	    dest[i*dim + k] = lrand48()/(float)INT_MAX;
+      }
+      n--;
+      count++;
+    }
+    return count;
+  }
+
+  virtual int ferror() { return 0; }
+  
+  virtual int feof() { return n <= 0; }
+
+private:
+  long n;
+};
+
+class FileStream : public PStream 
+{
+public:
+  FileStream( char* filename ) 
+  {
+    fp = fopen( filename, "rb");
+    if( fp == NULL ) {
+      fprintf(stderr,"error opening file %s\n.",filename);
+      exit(1);
+    }
+  }
+  size_t read( float* dest, int dim, int num ) {
+    return std::fread(dest, sizeof(float)*dim, num, fp); 
+  }
+  int ferror() {
+    return std::ferror(fp);
+  }
+  int feof() {
+    return std::feof(fp);
+  }
+  ~FileStream() {
+    fprintf(stderr,"closing file stream\n");
+    fclose(fp);
+  }
+private:
+  FILE* fp;
+};
+
+/* compute Euclidean distance squared between two points */
+inline float dist( const Point &p1, const Point &p2, const int dim )
+{
+  float result=0.0;
+  for ( auto i( 0 ); i<dim; i++ )
+  {
+    result += ( p1.coord[i] - p2.coord[i] ) * ( p1.coord[i] - p2.coord[i] );
+  }
+  return(result);
+}
+
+#endif
